+++ packages/react-server/src/forks/ReactServerFormatConfig.dom-relay.js
@@ -7,4 +7,4 @@
  * @flow
  */
 
-export * from '../ReactDOMServerFormatConfig';
+export * from 'react-dom/src/server/ReactDOMServerFormatConfig';
